+++ contracts/multihop/src/error.rs
@@ -4,11 +4,11 @@ use soroban_sdk::contracterror;
 #[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
 #[repr(u32)]
 pub enum ContractError {
-    AlreadyInitialized = 100,
-    OperationsEmpty = 101,
-    IncorrectAssetSwap = 102,
-    AdminNotSet = 103,
-    SameAdmin = 104,
-    NoAdminChangeInPlace = 105,
-    AdminChangeExpired = 106,
+    AlreadyInitialized = 200,
+    OperationsEmpty = 201,
+    IncorrectAssetSwap = 202,
+    AdminNotSet = 203,
+    SameAdmin = 204,
+    NoAdminChangeInPlace = 205,
+    AdminChangeExpired = 206,
 }